# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: SSCMS API
  description: 'Smart Student Club Management System -- Backend Sprint 1 (Milestone
    3) covered Epics 1-3: Role-Based Portal Access & Onboarding, Venue Booking & Asset
    Management, and Event Lifecycle & Approval Workflows. Backend Sprint 2 (Milestone
    4) adds Epics 4-6: Task Management & Volunteer Workflows, QR Attendance & Live
    Execution, and Certification & Budget Analytics.'
  version: 1.1.0
paths:
...
  /events/{event_id}/registrations:
    post:
      tags:
      - QR Attendance
      summary: Register For Event
      operationId: register_for_event_events__event_id__registrations_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationOut'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: Already registered for this event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - QR Attendance
      summary: List Event Registrations
      operationId: list_event_registrations_events__event_id__registrations_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationOut'
                title: Response List Event Registrations Events  Event Id  Registrations
                  Get
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /registrations/mine:
    get:
      tags:
      - QR Attendance
      summary: List My Registrations
      operationId: list_my_registrations_registrations_mine_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RegistrationOut'
                type: array
                title: Response List My Registrations Registrations Mine Get
      security:
      - OAuth2PasswordBearer: []
  /registrations/{qr_token}/check-in:
    post:
      tags:
      - QR Attendance
      summary: Check In
      operationId: check_in_registrations__qr_token__check_in_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: qr_token
        in: path
        required: true
        schema:
          type: string
          title: Qr Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationOut'
        '403':
          description: Caller cannot check attendees in for this event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Registration not found for this token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: Already checked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /events/{event_id}/attendance-stats:
    get:
      tags:
      - QR Attendance
      summary: Get Attendance Stats
      description: 'Return real-time attendance statistics for an event.


        Accessible to the event''s Club Head, Club President, or Faculty

        Coordinator -- the same roles that can view the full roster.'
      operationId: get_attendance_stats_events__event_id__attendance_stats_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceStatsOut'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AttendanceStatsOut:
      properties:
        event_id:
          type: integer
          title: Event Id
        total_registered:
          type: integer
          title: Total Registered
        checked_in:
          type: integer
          title: Checked In
        pending:
          type: integer
          title: Pending
        last_check_in_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Check In At
      type: object
      required:
      - event_id
      - total_registered
      - checked_in
      - pending
      title: AttendanceStatsOut
...
    ErrorOut:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RegistrationOut:
      properties:
        id:
          type: integer
          title: Id
        event_id:
          type: integer
          title: Event Id
        user_id:
          type: integer
          title: User Id
        qr_token:
          type: string
          title: Qr Token
        status:
          type: string
          title: Status
        registered_at:
          type: string
          format: date-time
          title: Registered At
        checked_in_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Checked In At
      type: object
      required:
      - id
      - event_id
      - user_id
      - qr_token
      - status
      - registered_at
      title: RegistrationOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login